Splits up project into smaller files.
- Moved configuration and global constants into thermo_config.h

- Changed some int types to byte to squeeze in a byte here and there.

- Changed display_err() to display_msg() so we can use this for
  reporting status changes.  Also added bootup messages so we
  get some feedback on what is going on.

- Changed serial output a bit for the values so they are easier to cut
  and paste into files to plot data.  (Not really useful, but we'll
  get back to that)

+++ thermo/networking.cpp
@@ -20,7 +20,7 @@ static char pass[] = WIFI_PASS;
 static int status = WL_IDLE_STATUS;
 
 bool connect_to_wifi() {
-    int retry = 0;
+    byte retry = 0;
 
     Serial.println("[WIFI] Initializing...");
     while (retry < WIFI_MAX_RETRY) {
